--- client/app/contributors/components/social-link.tsx
@@ -1,8 +1,0 @@
-import React from "react"
-
-interface ISocialLinkProps {}
-const SocialLink = () => {
-  return <div>SocialLink</div>
-}
-
-export default SocialLink
