+++ client/modules/contributors/components/contribution-area-badge.tsx
@@ -1,8 +1,8 @@
 import React from "react"
+import { ContributedArea } from "@/modules/contributors/types/ contributed-area"
+import getContributionAreaText from "@/modules/contributors/uilts/get-contribution-area-text"
 
 import { Badge } from "@/components/ui/badge"
-import { ContributedArea } from "@/app/contributors/types/ contributed-area"
-import getContributionAreaText from "@/app/contributors/uilts/get-contribution-area-text"
 
 interface IContributionAreaBadgeProps {
   area: ContributedArea
