+++ client/modules/contributors/data/contributors.ts
@@ -1,0 +1,21 @@
+import { IContributor } from "@/modules/contributors/types/contributor"
+
+const contributors: IContributor[] = [
+  {
+    name: "Paul Wong (aka R1R69)",
+    links: {
+      github: "https://github.com/Pjaijai",
+      linkedin: "https://www.linkedin.com/in/paulwong169/",
+    },
+    contributedArea: ["software_development", "uiux_design"],
+  },
+  {
+    name: "Gabby Chan",
+    links: {
+      linkedin: "https://www.linkedin.com/in/gabby-chan-94720b179/",
+    },
+    contributedArea: ["marketing", "graphic_design"],
+  },
+]
+
+export default contributors
