+++ client/modules/contributors/types/ contributed-area.ts
@@ -2,3 +2,4 @@ export type ContributedArea =
   | "software_development"
   | "uiux_design"
   | "marketing"
+  | "graphic_design"
